判断有向图是否存在有向环

+++ src/Graph/base/Edge.java
@@ -1,13 +1,11 @@
 package Graph.base;
 
-import Graph.Graph;
-
 public class Edge {
     public int u;//起点
     public int v;//终点
-    public int w;//边的权值
+    public double w;//边的权值
 
-    public Edge(int u, int v, int w) {
+    public Edge(int u, int v, double w) {
         this.u = u;
         this.v = v;
         this.w = w;
